move protocol to separate module
