+++ distributed/client.py
@@ -34,7 +34,7 @@ from .utils_comm import WrappedKey, unpack_remotedata, pack_data
 from .compatibility import Queue as pyQueue, Empty, isqueue
 from .core import (read, write, connect, coerce_to_rpc, clean_exception)
 from .protocol import to_serialize
-from .serialize.pickle import dumps, loads
+from .protocol.pickle import dumps, loads
 from .worker import dumps_function, dumps_task
 from .utils import (All, sync, funcname, ignoring, queue_to_iterator,
         tokey, log_errors, str_graph)
