+++ distributed/protocol/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import print_function, division, absolute_import
+
+from .compression import compressions, default_compression
+from .core import dumps, loads, maybe_compress, decompress, msgpack
+from .serialize import (serialize, deserialize, Serialize, Serialized,
+    to_serialize, register_serialization)
+
+from . import numpy
