+++ distributed/protocol/pickle.py
