--- distributed/serialize/__init__.py
@@ -1,6 +1,0 @@
-from .core import serialize, deserialize, Serialize, Serialized
-
-from ..utils import ignoring
-
-with ignoring(ImportError):
-    from . import numpy
