+++ distributed/utils_comm.py
@@ -16,7 +16,7 @@ from toolz import merge, concat, groupby, drop, valmap
 
 from .core import rpc, coerce_to_rpc, coerce_to_address
 from .utils import ignore_exceptions, All, log_errors, tokey, sync
-from .serialize.pickle import dumps, loads
+from .protocol.pickle import dumps, loads
 
 
 no_default = '__no_default__'
