+++ setup.py
@@ -22,14 +22,14 @@ setup(name='distributed',
       include_package_data=True,
       install_requires=requires,
       packages=['distributed',
-                'distributed.cli',
-                'distributed.diagnostics',
                 'distributed.bokeh',
                 'distributed.bokeh.status',
                 'distributed.bokeh.tasks',
                 'distributed.bokeh.workers',
+                'distributed.cli',
                 'distributed.deploy',
-                'distributed.serialize',
+                'distributed.diagnostics',
+                'distributed.protocol',
                 'distributed.http'],
       long_description=(open('README.md').read() if os.path.exists('README.md')
                         else ''),
